Crate cpio

A library for reading and writing CPIO archives.

CPIO archives can be in any of several formats. For now, this library only supports the newc (SVR4) format.

Re-exports

pub use newc::Builder as NewcBuilder;

pub use newc::Reader as NewcReader;

Modules

newc

Read/write newc (SVR4) format archives.

Functions

write_cpio

Creates a new CPIO archive.