cpio::newc

Struct Builder

pub struct Builder { /* private fields */ }

Builds metadata for one entry to be written into an archive.

Implementations

impl Builder

pub fn new(name: &str) -> Self

Create the metadata for one CPIO entry

pub fn ino(self, ino: u32) -> Self

Set the inode number for this file. In modern times however, typically this is just a a unique index ID for the file, rather than the actual inode number.

pub fn mode(self, mode: u32) -> Self

Set the file’s “mode” - the same as an inode “mode” field - containing permission bits and a bit of metadata about the type of file represented.

pub fn uid(self, uid: u32) -> Self

Set this file’s UID.

pub fn gid(self, gid: u32) -> Self

Set this file’s GID.

pub fn nlink(self, nlink: u32) -> Self

Set the number of links associated with this file.

pub fn mtime(self, mtime: u32) -> Self

Set the modification time of this file.

pub fn dev_major(self, dev_major: u32) -> Self

Set the major component of the device ID, describing the device on which this file resides.

Device IDs are comprised of a major and minor component. The major component identifies the class of device, while the minor component identifies a specific device of that class.

pub fn dev_minor(self, dev_minor: u32) -> Self

Set the minor component of the device ID, describing the device on which this file resides.

Device IDs are comprised of a major and minor component. The major component identifies the class of device, while the minor component identifies a specific device of that class.

pub fn rdev_major(self, rdev_major: u32) -> Self

Set the major component of the rdev ID, describes the device that this file (inode) represents.

Device IDs are comprised of a major and minor component. The major component identifies the class of device, while the minor component identifies a specific device of that class.

pub fn rdev_minor(self, rdev_minor: u32) -> Self

Set the minor component of the rdev ID, field describes the device that this file (inode) represents.

Device IDs are comprised of a major and minor component. The major component identifies the class of device, while the minor component identifies a specific device of that class.

pub fn set_mode_file_type(self, file_type: ModeFileType) -> Self

Set the mode file type of the entry

pub fn write<W: Write>(self, w: W, file_size: u32) -> Writer<W>

Write out an entry to the provided writer in SVR4 “new ascii” CPIO format.

pub fn write_crc<W: Write>( self, w: W, file_size: u32, file_checksum: u32, ) -> Writer<W>

Write out an entry to the provided writer in SVR4 “new crc” CPIO format.